+++ app/sprinkles/core/src/Bakery/MigrateRollbackCommand.php
@@ -11,7 +11,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
-use UserFrosting\System\Bakery\BaseCommand;
+use UserFrosting\Sprinkle\Core\Bakery\MigrateCommand;
 use UserFrosting\System\Bakery\Migrator;
 
 /**
@@ -20,7 +20,7 @@ use UserFrosting\System\Bakery\Migrator;
  *
  * @author Alex Weissman (https://alexanderweissman.com)
  */
-class MigrateRollback extends BaseCommand
+class MigrateRollbackCommand extends MigrateCommand
 {
     /**
      * {@inheritDoc}
@@ -29,8 +29,8 @@ class MigrateRollback extends BaseCommand
     {
         $this->setName("migrate:rollback")
              ->setDescription("Rollback last database migration")
-             ->addOption('steps', 's', InputOption::VALUE_REQUIRED, 'Number of steps to rollback', 1)
-             ->addOption('sprinkle', null, InputOption::VALUE_REQUIRED, 'The sprinkle to rollback', "")
+             ->addOption('steps', 's', InputOption::VALUE_REQUIRED, 'Number of batch to rollback', 1)
+             ->addOption('database', 'd', InputOption::VALUE_REQUIRED, 'The database connection to use')
              ->addOption('pretend', 'p', InputOption::VALUE_NONE, 'Run migrations in "dry run" mode');
     }
 
@@ -41,11 +41,22 @@ class MigrateRollback extends BaseCommand
     {
         $this->io->title("Migration rollback");
 
-        $step = $input->getOption('steps');
-        $sprinkle = $input->getOption('sprinkle');
+        $steps = $input->getOption('steps');
         $pretend = $input->getOption('pretend');
 
-        $migrator = new Migrator($this->io, $this->ci);
-        $migrator->runDown($step, $sprinkle, $pretend);
+        // Rollback migrations
+        $migrator = $this->setupMigrator($input);
+        $migrated = $migrator->rollback(['pretend' => $pretend, 'steps' => $steps]);
+
+        // Get notes and display them
+        $this->io->writeln($migrator->getNotes());
+
+        // If all went well, there's no fatal errors and we have migrated
+        // something, show some success
+        if (empty($migrated)) {
+            $this->io->success("Nothing to rollback");
+        } else {
+            $this->io->success("Rollback successful !");
+        }
     }
 }